+++ src/fides/api/service/saas_request/override_implementations/braze_request_overrides.py
@@ -12,8 +12,8 @@ from fides.api.service.saas_request.saas_request_override_factory import (
 from fides.api.util.saas_util import PRIVACY_REQUEST_ID
 
 
-@register("domo_user_update", [SaaSRequestType.UPDATE])
-def domo_user_update(
+@register("braze_user_update", [SaaSRequestType.UPDATE])
+def braze_user_update(
     client: AuthenticatedClient,
     param_values_per_row: List[Dict[str, Any]],
     policy: Policy,
@@ -23,23 +23,22 @@ def domo_user_update(
     rows_updated = 0
     # each update_params dict correspond to a record that needs to be updated
     for row_param_values in param_values_per_row:
-        # get params to be used in update request
-        user_id = row_param_values.get("user_id")
-
         # check if the privacy_request targeted emails for erasure,
-        # if so rewrite with a format that can be accepted by Domo
+        # if so rewrite with a format that can be accepted by Braze
         # regardless of the masking strategy in use
         all_object_fields = row_param_values["all_object_fields"]
 
         if "email" in all_object_fields:
             privacy_request_id = row_param_values[PRIVACY_REQUEST_ID]
             all_object_fields["email"] = f"{privacy_request_id}@company.com"
-            all_object_fields["alternateEmail"] = f"{privacy_request_id}@company.com"
 
-        update_body = dumps(all_object_fields)
+        update_body = dumps({"attributes": [all_object_fields]})
         client.send(
             SaaSRequestParams(
-                method=HTTPMethod.PUT, path=f"/v1/users/{user_id}", body=update_body
+                method=HTTPMethod.POST,
+                headers={"Content-Type": "application/json"},
+                path="/users/track",
+                body=update_body,
             )
         )
         rows_updated += 1